reorganize files and rename
