--- tests/patch_tracking/__init__.py
@@ -1,2 +1,0 @@
-from .task import *
-from .solver import *
